refactor/#38: 컴포넌트 import 경로 수정
